feat: um crime em um commit: Autenticaçao,Autorizaçao, corrige update endpoint, atualiza git ignore, outros

+++ src/GeoPet/Services/OwnerService.cs
@@ -1,37 +1,47 @@
 using AutoMapper;
-using GeoPet.DataContract.Model;
 using GeoPet.DataContract.Request;
 using GeoPet.DataContract.Response;
+using GeoPet.DataContract.Model;
 using GeoPet.Interfaces.Repository;
 using GeoPet.Interfaces.Services;
 using GeoPet.Validation;
 using GeoPet.Validation.Base;
-using Microsoft.Data.SqlClient;
-using Microsoft.EntityFrameworkCore;
+
 
 namespace GeoPet.Service;
 
-public class PetParentService : IUserService
+public class OwnerService : IOwnerService
 {
-    private readonly IUserRepository _petParentRepository;
+    private readonly IOwnerRepository _petParentRepository;
     private readonly IMapper _mapper;
     private readonly ISecurityServices _securityServices;
     private readonly ICepService _cepService;
+    private readonly IHttpContextAccessor _httpContextAcessor;
 
 
-    public PetParentService(IUserRepository petParentRepository, IMapper mapper, ISecurityServices securityServices, ICepService cepService)
+    public OwnerService(IOwnerRepository petParentRepository, IMapper mapper, ISecurityServices securityServices, ICepService cepService, IHttpContextAccessor httpContextAccessor)
     {
         _petParentRepository = petParentRepository;
         _mapper = mapper;
         _securityServices = securityServices;
         _cepService = cepService;
+        _httpContextAcessor = httpContextAccessor;
     }
 
-    public async Task<Response> CreateUser(UserRequest userRequest)
+    private bool ValidateAuthorization(int id)
+    {
+        var claim = int.Parse(_httpContextAcessor.HttpContext?.User.Claims.FirstOrDefault(c => c.Type == "userId")?.Value!);
+
+        if (claim != 1 && claim != id) return true;
+
+        return false;
+    }
+
+    public async Task<Response> CreateUser(OwnerRequest userRequest)
     {
         try
         {
-            var petParentValidation = new PetParentValidate(_cepService);
+            var petParentValidation = new OwnerValidate(_cepService);
             var petIsValid = petParentValidation.Validate(userRequest);
 
             var errors = GetValidations.GetErrors(petIsValid);
@@ -39,16 +49,11 @@ public class PetParentService : IUserService
             if (errors.Report.Any())
                 return errors;
 
-            // var isEquals = await _securityServices.ComparePassword(petParentRequest.Password, petParentRequest.ConfirmPassword);
-
-            // if (!isEquals.Data)
-            //     return Response.Unprocessable(Report.Create("Os password não são iguais."));
-
             var passwordEncripted = await _securityServices.EncryptPassword(userRequest.Password);
 
-            userRequest.Password = passwordEncripted.Data;
+            userRequest.Password = passwordEncripted.Data!;
 
-            var petParent = _mapper.Map<User>(userRequest);
+            var petParent = _mapper.Map<DataContract.Model.Owner>(userRequest);
 
             var petParentAdd = _petParentRepository.Add(petParent);
 
@@ -68,7 +73,7 @@ public class PetParentService : IUserService
         try
         {
             var petParents = await _petParentRepository.GetAll();
-            var petParentResponseList = _mapper.Map<List<User>, List<UserResponse>>(petParents);
+            var petParentResponseList = _mapper.Map<List<DataContract.Model.Owner>, List<UserResponse>>(petParents);
             var response = new Response<List<UserResponse>>(petParentResponseList);
             return response;
         }
@@ -82,6 +87,8 @@ public class PetParentService : IUserService
     {
         try
         {
+            if (ValidateAuthorization(id)) return Response.Unprocessable(Report.Create("Sem permissão"));
+
             var petParent = await _petParentRepository.GetById(id);
             var petParentResponse = _mapper.Map<UserResponse>(petParent);
             var response = new Response<UserResponse>(petParentResponse);
@@ -93,23 +100,29 @@ public class PetParentService : IUserService
         }
     }
 
-    public async Task<Response> UpdateUser(int id, UserRequest petParentRequest)
+    public async Task<Response> UpdateUser(int id, OwnerRequest ownerRequest)
     {
         try
         {
-            var petParentValidation = new PetParentValidate(_cepService);
-            var petIsValid = petParentValidation.Validate(petParentRequest);
+            if (ValidateAuthorization(id)) return Response.Unprocessable(Report.Create("Sem permissão"));
+
+            var petParentValidation = new OwnerValidate(_cepService);
+            var petIsValid = petParentValidation.Validate(ownerRequest);
 
             var errors = GetValidations.GetErrors(petIsValid);
 
             if (errors.Report.Any())
                 return errors;
 
-            var petParent = _mapper.Map<User>(petParentRequest);
+            var passwordEncripted = await _securityServices.EncryptPassword(ownerRequest.Password);
+            ownerRequest.Password = passwordEncripted.Data!;
 
-            var petParentUpdated = _petParentRepository.Update(petParent);
+            var petParent = _mapper.Map<Owner>(ownerRequest);
+            petParent.Id = id;
 
-            var petParentResponse = _mapper.Map<UserResponse>(petParentUpdated.Result);
+            var petParentUpdated = await _petParentRepository.Update(id, petParent);
+
+            var petParentResponse = _mapper.Map<UserResponse>(petParentUpdated);
 
             var response = new Response<UserResponse>(petParentResponse);
             return response;
@@ -124,6 +137,8 @@ public class PetParentService : IUserService
     {
         try
         {
+            if (ValidateAuthorization(id)) return Response.Unprocessable(Report.Create("Sem permissão"));
+
             var petParentToDelete = _petParentRepository.GetById(id);
 
             var petParentDeleted = await _petParentRepository.Remove(petParentToDelete.Result);
